ns: change PNP testing to use QNNPACK
Summary:

Changes the PNP test cases to use QNNPACK. The only reason is because
I'm switching to Mac M1 as my primary machine, which supports QNNPACK
but not fbgemm, and it's convenient for me to be able to run these
locally.

PNP itself is not backend specific, so it does not matter which backend
the functionality is tested on.

Test plan:

```
python test/test_quantization.py -k NShadows
```

ghstack-source-id: 8c560052d63d50451d2bd5e8acf77442a8dbfa21
Pull Request resolved: https://github.com/pytorch/pytorch/pull/91421